Game Mostly Works
Location Activity pops up when you reach a location. Play button calls the Main Activity but must delay by 1 second or else things are messed up when drawing the assets in the game. In the game, Arte can move and hit the assets, win the game, and pop up the activity based on what place you are.