Fixed Buttons but Probably Need to Remove Them
Fixed the Main Menu buttons on each locaiton activity, but they will probably need to be removed, as creating a new instance of the Entry Activity appears to check the GPS location again and immediatley pop up with the activity again instead of the Entry Activity. This looks broken to a user. Using the back button already present on an Android device does not trigger this and would be best.